requirements.txt
python-dotenv
streamlit
langchain
langchain-community
langchain-openai
faiss-cpu
